fix: Full update entity - Update all entity - Add audit config - Update Categorize enum - Comment out functions that cannot be run

+++ src/main/java/com/swd392/ticket_resell_be/services/impls/VNPayServiceImplement.java
@@ -14,7 +14,7 @@ import java.util.*;
 @Service
 public class VNPayServiceImplement {
 
-    public VNPayOrderResponse createOrder(int total, String orderInfor){
+    public VNPayOrderResponse createOrder(float total, String orderInfor) {
         String vnp_Version = "2.1.0";
         String vnp_Command = "pay";
         String vnp_TxnRef = VNPayConfig.getRandomNumber(8);
@@ -26,7 +26,7 @@ public class VNPayServiceImplement {
         vnp_Params.put("vnp_Version", vnp_Version);
         vnp_Params.put("vnp_Command", vnp_Command);
         vnp_Params.put("vnp_TmnCode", vnp_TmnCode);
-        vnp_Params.put("vnp_Amount", String.valueOf(total*100));
+        vnp_Params.put("vnp_Amount", String.valueOf(total * 100));
         vnp_Params.put("vnp_CurrCode", "VND");
 
         vnp_Params.put("vnp_TxnRef", vnp_TxnRef);
@@ -83,9 +83,9 @@ public class VNPayServiceImplement {
         return new VNPayOrderResponse(paymentUrl, vnp_TxnRef);
     }
 
-    public int orderReturn(HttpServletRequest request){
+    public int orderReturn(HttpServletRequest request) {
         Map fields = new HashMap();
-        for (Enumeration params = request.getParameterNames(); params.hasMoreElements();) {
+        for (Enumeration params = request.getParameterNames(); params.hasMoreElements(); ) {
             String fieldName = null;
             String fieldValue = null;
             try {
